BAH-4876 | Move SearchResponse from search/ to search/criteria/ (response DTO)
